Correccion tipo de dato en Productos

+++ src/main/java/co/edu/poli/tiendadj/modelo/ProductoAlimenticio.java
@@ -1,18 +1,18 @@
 package co.edu.poli.tiendadj.modelo;
 
 public class ProductoAlimenticio extends Producto {
-    private float aporteCalorico;
+    private double aporteCalorico;
 
-    public ProductoAlimenticio(int id, String descripcion, float aporteCalorico) {
+    public ProductoAlimenticio(int id, String descripcion, double aporteCalorico) {
         super(id, descripcion);
         this.aporteCalorico = aporteCalorico;
     }
 
-    public float getAporteCalorico() {
+    public double getAporteCalorico() {
         return aporteCalorico;
     }
 
-    public void setAporteCalorico(float aporteCalorico) {
+    public void setAporteCalorico(double aporteCalorico) {
         this.aporteCalorico = aporteCalorico;
     }
 }
