+++ src/main/java/co/edu/poli/tiendadj/modelo/ProductoElectronico.java
@@ -1,17 +1,17 @@
 package co.edu.poli.tiendadj.modelo;
 
 public class ProductoElectronico extends Producto {
-    private float voltInput;
-    public ProductoElectronico(int id, String descripcion, float voltInput) {
+    private double voltInput;
+    public ProductoElectronico(int id, String descripcion, double voltInput) {
         super(id, descripcion);
         this.voltInput = voltInput;
     }
 
-    public float getVoltInput() {
+    public double getVoltInput() {
         return voltInput;
     }
 
-    public void setVoltInput(float voltInput) {
+    public void setVoltInput(double voltInput) {
         this.voltInput = voltInput;
     }
 }
